break bookings and sessions controllers into different specs
